ewm-service/src/main/java/ru/practicum/explore/ewm/users/UserController.java
package ru.practicum.explore.ewm.users;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import ru.practicum.explore.ewm.users.dto.UserDto;
import ru.practicum.explore.ewm.users.dto.UserMapper;
import ru.practicum.explore.ewm.users.dto.UserNewDto;

import javax.validation.Valid;
import javax.validation.constraints.Positive;
import javax.validation.constraints.PositiveOrZero;
import java.util.List;
import java.util.stream.Collectors;

@RestController
@RequestMapping("/admin/users")
@Validated
@RequiredArgsConstructor
@Slf4j
public class UserController {
    private final UserService userService;
    private final UserMapper userMapper;

    @GetMapping
    public List<UserDto> getUsers(@RequestParam(required = false) List<Long> ids,
                                  @RequestParam(defaultValue = "0") @PositiveOrZero int from,
                                  @RequestParam(defaultValue = "10") @Positive int size) {
        log.info("getUsers " + ids);
        return userService.getUsers(ids, from, size)
                .stream()
                .map(userMapper::fromUserToDto)
                .collect(Collectors.toList());
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public UserDto addUser(@RequestBody @Valid UserNewDto newUser) {
        return userMapper.fromUserToDto(userService.addUser(userMapper.fromDtoNewToUser(newUser)));
    }

    @DeleteMapping("/{userId}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@PathVariable Long userId) {
        userService.findById(userId);
        userService.delete(userId);
    }

}
